# lib/api-zod/src/generated/types/categoryRef.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryLevel } from './categoryLevel';

export interface CategoryRef {
  dimension: CategoryLevel;
  theme: CategoryLevel;
  direction: CategoryLevel;
}