# lib/api-zod/src/generated/types/materialAnalysis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryRef } from './categoryRef';

export interface MaterialAnalysis {
  title: string;
  summary: string;
  primary: CategoryRef;
  /** @maxItems 3 */
  secondary: CategoryRef[];
  /** @maxItems 5 */
  keywords: string[];
  /** @maxItems 3 */
  argument_types: string[];
  core_value: string;
}